# server/docs/swagger.yaml
definitions:
...
  nfc_relay_admin.ConfigAuditResponse:
    properties:
      page:
        type: integer
      records:
        items:
          $ref: '#/definitions/nfc_relay_admin.ConfigChangeRecord'
        type: array
      size:
        type: integer
      total:
        type: integer
    type: object
  nfc_relay_admin.ConfigAuditStats:
    properties:
      changesByConfig:
        additionalProperties:
          type: integer
        type: object
      changesByOperator:
        additionalProperties:
          type: integer
        type: object
      changesByType:
        additionalProperties:
          type: integer
        type: object
      changesThisMonth:
        type: integer
      changesThisWeek:
        type: integer
      changesToday:
        type: integer
      configChangesTrend:
        items:
          $ref: '#/definitions/nfc_relay_admin.ConfigChangeTrendPoint'
        type: array
      recentOperators:
        items:
          $ref: '#/definitions/nfc_relay_admin.OperatorActivity'
        type: array
      successRate:
        type: number
      totalChanges:
        type: integer
    type: object
  nfc_relay_admin.ConfigChangeRecord:
    properties:
      changeType:
        description: CREATE, UPDATE, DELETE, RELOAD
        type: string
      changedFields:
        items:
          type: string
        type: array
      configType:
        type: string
      errorMessage:
        type: string
      id:
        type: integer
      ipAddress:
        type: string
      newValue:
        additionalProperties: true
        type: object
      oldValue:
        additionalProperties: true
        type: object
      operatorId:
        type: string
      operatorName:
        type: string
      reason:
        type: string
      rollbackId:
        type: integer
      sessionId:
        type: string
      status:
        description: SUCCESS, FAILED, PENDING
        type: string
      timestamp:
        type: string
      userAgent:
        type: string
    type: object
  nfc_relay_admin.ConfigChangeTrendPoint:
    properties:
      changeCount:
        type: integer
      date:
        type: string
      failedCount:
        type: integer
      successCount:
        type: integer
    type: object
...
  nfc_relay_admin.OperatorActivity:
    properties:
      changeCount:
        type: integer
      lastActivity:
        type: string
      operatorId:
        type: string
      operatorName:
        type: string
      successRate:
        type: number
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.2
paths:
...
  /admin/nfc-relay/v1/config-audit/changes/{change_id}:
    get:
      consumes:
      - application/json
      description: 获取指定配置变更记录的详细信息
      parameters:
      - description: 变更记录ID
        in: path
        name: change_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/nfc_relay_admin.ConfigChangeRecord'
              type: object
      summary: 获取配置变更详情
      tags:
      - NFC中继管理
  /admin/nfc-relay/v1/config-audit/export:
    get:
      consumes:
      - application/json
      description: 导出配置审计日志到文件
      parameters:
      - description: 导出格式
        enum:
        - csv
        - xlsx
        - json
        in: query
        name: format
        type: string
      - description: 配置类型
        in: query
        name: configType
        type: string
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 结束时间
        in: query
        name: endTime
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: 导出配置审计日志
      tags:
      - NFC中继管理
  /admin/nfc-relay/v1/config-audit/logs:
    get:
      consumes:
      - application/json
      description: 查询配置变更的审计日志记录
      parameters:
      - description: 配置类型
        in: query
        name: configType
        type: string
      - description: 变更类型
        in: query
        name: changeType
        type: string
      - description: 操作员ID
        in: query
        name: operatorId
        type: string
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 结束时间
        in: query
        name: endTime
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/nfc_relay_admin.ConfigAuditResponse'
              type: object
      summary: 获取配置审计日志
      tags:
      - NFC中继管理
  /admin/nfc-relay/v1/config-audit/records:
    post:
      consumes:
      - application/json
      description: 手动创建配置变更审计记录
      parameters:
      - description: 配置变更记录
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/nfc_relay_admin.ConfigChangeRecord'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建配置审计记录
      tags:
      - NFC中继管理
  /admin/nfc-relay/v1/config-audit/stats:
    get:
      consumes:
      - application/json
      description: 获取配置变更的统计信息和趋势
      parameters:
      - description: 统计天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/nfc_relay_admin.ConfigAuditStats'
              type: object
      summary: 获取配置审计统计
      tags:
      - NFC中继管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...